Javascript code that does not stink
A bit of refactoring to make something more appropiate

+++ src/main/resources/static/js/show.js
@@ -3,11 +3,25 @@ function showShoppingList(id) {
     //take data
     $.getJSON('shoppinglist/'+id, function (data) {
         emptyWindow();
-        showList(data);
+        var listToShow = new ShowList(data);
+        listToShow.show();
     });
 }
 
-function createElement(product) {
+function ShowList(data) {
+    console.log("Ready to write data " + JSON.stringify(data));
+    this.form = getTemplate('#listShowFormTemplate');
+    this.form.querySelector('h3').innerHTML = data.name;
+    this.form.querySelector('#deleteList').addEventListener('click', function() { deleteList(data.id); });
+    this.form.querySelector('#editList').addEventListener('click', function() { editList(data); });
+    var objThis = this;
+    data.products.forEach(function(product) {
+        objThis.createElement(product);
+    });
+    console.log("Appending child: " + this.form.firstElementChild.outerHTML);
+}
+
+ShowList.prototype.createElement = function(product) {
     var myElement = getTemplate('#listElementShowFormTemplate');
     var input = myElement.querySelector('input');
     var label = myElement.querySelector('label');
@@ -15,44 +29,31 @@ function createElement(product) {
     input.checked = product.mark;
     label.setAttribute('for', 'product-'+product.id);
     label.innerHTML = product.mark? '<del>'+product.name+'</del>' : product.name;
-    return myElement;
+    console.log("Appending: "+myElement.firstElementChild.outerHTML);
+    this.form.querySelector('form div').appendChild(myElement);
 }
 
-function showList(data) {
-    console.log("Ready to write data " + JSON.stringify(data));
-    var myForm = getTemplate('#listShowFormTemplate');
-    myForm.querySelector('h3').innerHTML = data.name;
-    myForm.querySelector('#deleteList').addEventListener('click', function() { deleteList(data.id); });
-    myForm.querySelector('#editList').addEventListener('click', function() { editList(data); });
-    data.products.forEach(function(product) {
-        var myElement = createElement(product);
-        console.log("Appending: "+myElement.firstElementChild.outerHTML);
-        myForm.querySelector('form div').appendChild(myElement);
-    });
-    console.log("Appending child: " + myForm.firstElementChild.outerHTML);
-    $('#listArea').append(myForm);
-}
-
-function editList(data) {
-    console.log('Editing list: '+JSON.stringify(data));
-    newList(data);
+ShowList.prototype.show = function() {
+    $('#listArea').append(this.form);
 }
 
 function deleteList(identifier) {
     console.log('Deleting list: '+identifier);
 
-    $.ajax({
-        type: "DELETE",
-        contentType: "application/json",
-        url: "/shoppinglist/"+identifier,
-        timeout: 600000,
-        success: function () {
-            console.log("DONE");
-            location.href = '/';
-        },
-        error: function (e) {
-            console.log("ERROR: ", e);
-            display(e);
-        }
-    });
+    if(confirm("¿Seguro que quiere borrar la lista?")) {
+        $.ajax({
+            type: "DELETE",
+            contentType: "application/json",
+            url: "/shoppinglist/" + identifier,
+            timeout: 600000,
+            success: function () {
+                console.log("DONE");
+                location.href = '/';
+            },
+            error: function (e) {
+                console.log("ERROR: ", e);
+                display(e);
+            }
+        });
+    }
 }
